+++ src/Pwa/GenerateIndexHtmls.mjs
@@ -1,39 +1,40 @@
+import { LOCALIZATION_MODULE_PWA_GENERATOR } from "../Localization/LOCALIZATION_MODULE.mjs";
 import { dirname, join } from "node:path/posix";
 import { readFile, writeFile } from "node:fs/promises";
 
-/** @typedef {import("../../../flux-localization-api/src/FluxLocalizationApi.mjs").FluxLocalizationApi} FluxLocalizationApi */
 /** @typedef {import("../FluxPwaGenerator.mjs").FluxPwaGenerator} FluxPwaGenerator */
+/** @typedef {import("../Localization/Localization.mjs").Localization} Localization */
 
 export class GenerateIndexHtmls {
-    /**
-     * @type {FluxLocalizationApi | null}
-     */
-    #flux_localization_api;
     /**
      * @type {FluxPwaGenerator}
      */
     #flux_pwa_generator;
+    /**
+     * @type {Localization | null}
+     */
+    #localization;
 
     /**
      * @param {FluxPwaGenerator} flux_pwa_generator
-     * @param {FluxLocalizationApi | null} flux_localization_api
+     * @param {Localization | null} localization
      * @returns {GenerateIndexHtmls}
      */
-    static new(flux_pwa_generator, flux_localization_api = null) {
+    static new(flux_pwa_generator, localization = null) {
         return new this(
             flux_pwa_generator,
-            flux_localization_api
+            localization
         );
     }
 
     /**
      * @param {FluxPwaGenerator} flux_pwa_generator
-     * @param {FluxLocalizationApi | null} flux_localization_api
+     * @param {Localization | null} localization
      * @private
      */
-    constructor(flux_pwa_generator, flux_localization_api) {
+    constructor(flux_pwa_generator, localization) {
         this.#flux_pwa_generator = flux_pwa_generator;
-        this.#flux_localization_api = flux_localization_api;
+        this.#localization = localization;
     }
 
     /**
@@ -45,12 +46,13 @@ export class GenerateIndexHtmls {
      */
     async generateIndexHtmls(index_template_html_file, index_html_file, manifest_json_file, localization_folder = null) {
         if (localization_folder !== null) {
-            if (this.#flux_localization_api === null) {
-                throw new Error("Missing FluxLocalizationApi");
+            if (this.#localization === null) {
+                throw new Error("Missing Localization");
             }
 
-            this.#flux_localization_api.addModule(
-                localization_folder
+            await this.#localization.addModule(
+                localization_folder,
+                LOCALIZATION_MODULE_PWA_GENERATOR
             );
         }
 
@@ -81,7 +83,9 @@ export class GenerateIndexHtmls {
         const web_manifest_json_file_dot_pos = web_manifest_json_file.lastIndexOf(".");
 
         for (const language of [
-            ...(localization_folder !== null ? Object.keys((await this.#flux_localization_api.getLanguages()).all) : null) ?? [],
+            ...(localization_folder !== null ? Object.keys((await this.#localization.getLanguages(
+                LOCALIZATION_MODULE_PWA_GENERATOR
+            )).all) : null) ?? [],
             ""
         ]) {
             const localized_index_html_file = language !== "" ? `${index_html_file.substring(0, index_html_file_dot_pos)}-${language}${index_html_file.substring(index_html_file_dot_pos)}` : index_html_file;